+++ kv/interceptor.go
@@ -9,8 +9,8 @@ import (
 // Interceptor defines functions that are invoked around keyspace operations.
 type Interceptor[K, V any] struct {
 	beforeOpen xatomic.Value[func(string) error]
-	beforeSet  xatomic.Value[func(string, K, V, *uint64) error]
-	afterSet   xatomic.Value[func(string, K, V, *uint64) error]
+	beforeSet  xatomic.Value[func(string, K, V, *Revision) error]
+	afterSet   xatomic.Value[func(string, K, V, *Revision) error]
 }
 
 // BeforeOpen sets the function that is invoked before a [Keyspace] is opened.
@@ -19,12 +19,12 @@ func (i *Interceptor[K, V]) BeforeOpen(fn func(name string) error) {
 }
 
 // BeforeSet sets the function that is invoked before a key/value pair is set.
-func (i *Interceptor[K, V]) BeforeSet(fn func(keyspace string, k K, v V, r *uint64) error) {
+func (i *Interceptor[K, V]) BeforeSet(fn func(keyspace string, k K, v V, r *Revision) error) {
 	i.beforeSet.Store(fn)
 }
 
 // AfterSet sets the function that is invoked after a key/value pair is set.
-func (i *Interceptor[K, V]) AfterSet(fn func(keyspace string, k K, v V, r *uint64) error) {
+func (i *Interceptor[K, V]) AfterSet(fn func(keyspace string, k K, v V, r *Revision) error) {
 	i.afterSet.Store(fn)
 }
 
@@ -75,7 +75,7 @@ func (ks *interceptedKeyspace[K, V]) Name() string {
 	return ks.Next.Name()
 }
 
-func (ks *interceptedKeyspace[K, V]) Get(ctx context.Context, k K) (V, uint64, error) {
+func (ks *interceptedKeyspace[K, V]) Get(ctx context.Context, k K) (V, Revision, error) {
 	return ks.Next.Get(ctx, k)
 }
 
@@ -83,7 +83,7 @@ func (ks *interceptedKeyspace[K, V]) Has(ctx context.Context, k K) (bool, error)
 	return ks.Next.Has(ctx, k)
 }
 
-func (ks *interceptedKeyspace[K, V]) Set(ctx context.Context, k K, v V, r uint64) error {
+func (ks *interceptedKeyspace[K, V]) Set(ctx context.Context, k K, v V, r Revision) error {
 	rClone := r
 
 	if fn := ks.Interceptor.beforeSet.Load(); fn != nil {